Add voice leveling, other changes

+++ src/main/java/union/utils/logs/LogEmbedUtil.java
@@ -1253,7 +1253,7 @@ public class LogEmbedUtil {
 
 	// Level
 	@NotNull
-	public MessageEmbed levelUp(DiscordLocale locale, Member member, long level) {
+	public MessageEmbed levelUp(DiscordLocale locale, Member member, int level) {
 		return new LogEmbedBuilder(locale, GREEN_DARK)
 			.setHeaderIcon(LogEvent.LEVEL_UP, member.getEffectiveAvatarUrl(), member.getUser().getName())
 			.setDescription(lu.getLocalizedRandom(locale, "logger.level.msg_random")
